+++ lib/src/types/boolean.dart
@@ -1,12 +1,20 @@
-part of bson;
+import '../../bson.dart';
 
 class BsonBoolean extends BsonObject {
   BsonBoolean(this.data);
   BsonBoolean.fromBuffer(BsonBinary buffer) : data = extractData(buffer);
+  BsonBoolean.fromEJson(ejson) : data = extractEJson(ejson);
 
   bool data;
 
   static bool extractData(BsonBinary buffer) => buffer.readByte() == 1;
+  static bool extractEJson(ejson) {
+    if (ejson is! bool) {
+      throw ArgumentError(
+          'The received Object is not a valid Bool representation');
+    }
+    return ejson;
+  }
 
   @override
   bool get value => data;
@@ -18,8 +26,5 @@ class BsonBoolean extends BsonObject {
   void packValue(BsonBinary buffer) => buffer.writeByte(data ? 1 : 0);
 
   @override
-  eJson({bool relaxed = false}) {
-    // TODO: implement eJson
-    throw UnimplementedError();
-  }
+  eJson({bool relaxed = false}) => data;
 }