+++ lib/src/types/string.dart
@@ -1,4 +1,8 @@
-part of bson;
+import 'dart:convert';
+
+import '../../bson.dart';
+import '../statics.dart';
+import '../utils/types_def.dart';
 
 class BsonString extends BsonObject {
   BsonString(this.data);