+++ lib/src/types/timestamp.dart
@@ -5,15 +5,36 @@ class BsonTimestamp extends BsonObject {
       : timestamp = parmTimestamp ??
             Timestamp((DateTime.now().millisecondsSinceEpoch ~/ 1000).toInt(),
                 Statics.nextIncrement);
-
   BsonTimestamp.fromBuffer(BsonBinary buffer) : timestamp = extractData(buffer);
-
+  BsonTimestamp.fromEJson(Map<String, dynamic> eJsonMap)
+      : timestamp = extractEJson(eJsonMap);
   Timestamp timestamp;
 
   static Timestamp extractData(BsonBinary buffer) {
     var increment = buffer.readInt32();
     var seconds = buffer.readInt32();
     return Timestamp(seconds, increment);
+  }
+
+  static Timestamp extractEJson(Map<String, dynamic> eJsonMap) {
+    var entry = eJsonMap.entries.first;
+    if (entry.key != type$timestamp) {
+      throw ArgumentError(
+          'The received Map is not a avalid EJson Timestamp representation');
+    }
+    if (entry.value is! Map<String, Object>) {
+      throw ArgumentError(
+          'The received Map is not a valid EJson Timestamp representation');
+    }
+    var content = entry.value as Map<String, Object>;
+    if (content.containsKey('t') && content.containsKey('i')) {
+      int seconds = content['t'] as int;
+      int increment = content['i'] as int;
+
+      return Timestamp(seconds, increment);
+    }
+    throw ArgumentError(
+        'The received Map is not a valid EJson Timestamp representation');
   }
 
   @override
@@ -31,8 +52,7 @@ class BsonTimestamp extends BsonObject {
   }
 
   @override
-  eJson({bool relaxed = false}) {
-    // TODO: implement eJson
-    throw UnimplementedError();
-  }
+  eJson({bool relaxed = false}) => {
+        type$timestamp: {'t': timestamp.seconds, 'i': timestamp.increment}
+      };
 }